feat: merged some files and add file of names
