+++ src/pages/feature-pages/tables/names_of_tables.json
@@ -1,0 +1,20 @@
+[
+    "Акушерство",
+    "ВАШ. НОШ. Шкалы оценки интенсивности боли",
+    "Критерии оценки новорождённого по шкале Апгар",
+    "Нормы ЧД, ЧСС, АД у детей (в покое)",
+    "Определение площади ожогов у детей (по Lund и Browder)",
+    "Острая дыхательная недостаточность (Кассиль В.Л. 2004 г.)",
+    "Оценка мышечной силы по баллам",
+    "Параметры проведения базовой СЛР",
+    "Проктокол оценки тяжести состояния пациента (NEWS)",
+    "Промывание желудка у детей",
+    "Размеры эндотрахеальных трубок у детей",
+    "Соответствие размеров ларннгеальных трубок параметрам пациента",
+    "ХСН ШОКС (в модификации Мареева В.Ю.)",
+    "Шкала Глазго (Glasgow Coma Scale)",
+    "Шкала возбуждения-седации Ричмонда (шкала RASS)",
+    "Шкала комы FOUR",
+    "Шкала моторного дефицита LAMS (Los Angeles Motor Scale)",
+    "Шкала оценки вероятности ТЭЛА (Revised Geneva Score)"
+  ]
